Change index in ITER object to be 4 bytes. Add wrapper for nrf24 powerdown.

+++ bitsy_python_arduino/datatypes/iter.h
@@ -6,7 +6,7 @@
 namespace bitsy_python {
 extern Variable IterCreate(const Variable& var);
 extern bool IterForLoopIter(const Variable& iter, Variable *elem);
-extern bool IterForLoop(Variable::CustomType iter, uint16_t ind, Variable *elem);
+extern bool IterForLoop(Variable::CustomType iter, uint32_t ind, Variable *elem);
 }
 
 #endif /* ITER_H__ */
